Update patch set 2
Patch Set 2:

> @Shawn do you know any quick way to apply a custom CSS to this Markdown rendering?

Our Markdown flavor doesn't support custom CSS or inline HTML. It is what it is.

> The simplest way I could envisage would have been having a different URL (e.g. www.gerritusersummit.com) with a NGINX reverse-proxy with an overridden local CSS.

Basically. It would have to be a new domain name, and a standalone web server that is serving the content. At that point it might as well just be running a fork of Gitiles, serving a clone of this repository or something. That just sounds like too much work to me for custom CSS.

Patch-set: 2
Reviewer: Gerrit User 1001240 <1001240@173816e5-2b9a-37c3-8a2e-48639d4f1153>